Se agregan vistas de productos y empleados, se genera el registro de productos y se comienza a generar el inicio de sesion

+++ src/app/api/products/[id]/route.js
@@ -3,50 +3,56 @@ import { pool } from "@/libs/mysql";
 
 export async function GET(request, { params }) {
   try {
+    // Validar que params.id sea un número
+    const id = parseInt(params.id, 10);
+    if (isNaN(id)) {
+      return NextResponse.json(
+        { message: "ID inválido. Debe ser un número." },
+        { status: 400 }
+      );
+    }
+
     const results = await pool.query(
       "SELECT * FROM INVENTARIO_MATERIAL WHERE ID = ?",
-      params.id
+      [id]
     );
 
     if (results.length === 0) {
       return NextResponse.json(
-        {
-          message: "Producto no encontrado",
-        },
-        {
-          status: 404,
-        }
+        { message: "Producto no encontrado" },
+        { status: 404 }
       );
     }
 
     return NextResponse.json(results[0]);
   } catch (error) {
-    return NextResponse.json(
-      {
-        message: error.message,
-      },
-      { status: 500 }
-    );
+    return NextResponse.json({ message: error.message }, { status: 500 });
   }
 }
 
 export async function DELETE(request, { params }) {
   try {
+    // Validar que params.id sea un número
+    const id = parseInt(params.id, 10);
+    if (isNaN(id)) {
+      return NextResponse.json(
+        { message: "ID inválido. Debe ser un número." },
+        { status: 400 }
+      );
+    }
+
     const results = await pool.query(
       "DELETE FROM INVENTARIO_MATERIAL WHERE ID = ?",
-      [params.id]
+      [id]
     );
 
     if (results.affectedRows === 0) {
       return NextResponse.json(
-        {
-          message: "Producto no encontrado",
-        },
+        { message: "Producto no encontrado" },
         { status: 404 }
       );
     }
 
-    // return new Response(null, { status: 204 });
     return NextResponse.json(null, { status: 204 });
   } catch (error) {
     return NextResponse.json({ message: error.message }, { status: 500 });
@@ -55,10 +61,49 @@ export async function DELETE(request, { params }) {
 
 export async function PUT(request, { params }) {
   try {
+    // Validar que params.id sea un número
+    const id = parseInt(params.id, 10);
+    if (isNaN(id)) {
+      return NextResponse.json(
+        { message: "ID inválido. Debe ser un número." },
+        { status: 400 }
+      );
+    }
+
     const data = await request.json();
-    const results = pool.query(
+
+    // Validar campos en el objeto `data`
+    const {
+      nombre,
+      descripcion,
+      categoria,
+      cantidad,
+      unidad_medida,
+      precio_unitario,
+    } = data;
+    if (
+      !nombre ||
+      !categoria ||
+      !cantidad ||
+      !unidad_medida ||
+      !precio_unitario ||
+      typeof nombre !== "string" ||
+      typeof categoria !== "string" ||
+      typeof unidad_medida !== "string" ||
+      typeof cantidad !== "number" ||
+      typeof precio_unitario !== "number" ||
+      cantidad < 0 ||
+      precio_unitario < 0
+    ) {
+      return NextResponse.json(
+        { message: "Datos inválidos o incompletos." },
+        { status: 400 }
+      );
+    }
+
+    const results = await pool.query(
       "UPDATE inventario_material SET ? WHERE ID = ?",
-      [data, params.id]
+      [data, id]
     );
 
     if (results.affectedRows === 0) {
@@ -70,9 +115,8 @@ export async function PUT(request, { params }) {
 
     const updatedProduct = await pool.query(
       "SELECT * FROM INVENTARIO_MATERIAL WHERE ID = ?",
-      [params.id]
+      [id]
     );
-    console.log(results);
 
     return NextResponse.json(updatedProduct[0]);
   } catch (error) {
